server/requirements.txt
flask
openai
chromadb

langgraph
langchain
langchain_core
langchain_community
langchain_openai
langchain_postgres
transformers
torch
python-docx
spacy
psycopg2
python-dotenv
flask_cors
flask_sqlalchemy
typing_extensions
uuid
